# API-Gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuthResponse:
    description: Ответ с данными аутентификации
    properties:
      role:
        example: ROLE_STUDENT
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  models.ErrorResponse:
    description: Расширенный ответ с ошибкой
    properties:
      code:
        example: INVALID_INPUT
        type: string
      details:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
...
  models.LoginRequest:
    description: Запрос на аутентификацию пользователя
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      role:
        enum:
        - ROLE_STUDENT
        - ROLE_DEVELOPER
        - ROLE_HR
        - ROLE_COMPANY
        example: ROLE_STUDENT
        type: string
    required:
    - email
    - password
    - role
    type: object
...
  models.SignUpRequest:
    description: Запрос на регистрацию нового пользователя
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      role:
        enum:
        - ROLE_STUDENT
        - ROLE_DEVELOPER
        - ROLE_HR
        - ROLE_COMPANY
        example: ROLE_STUDENT
        type: string
    required:
    - email
    - password
    - role
    type: object
...
  models.ValidationError:
    description: Детализированная ошибка валидации
    properties:
      code:
        example: REQUIRED_FIELD
        type: string
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@studjobs.ru
    name: API Support
  description: API Gateway для платформы StudJobs HH
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: StudJobs HH API Gateway
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя в систему и возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT токен
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RoleAuth:
    description: Роль пользователя
    in: header
    name: X-User-Role
    type: apiKey
swagger: "2.0"